refactor(notifications): split NotificationService into per-feature partials
NotificationService becomes a sealed partial class. Each cross-module contract
method (ListAsync, GetUnreadCountAsync, GetByIdAsync, MarkReadAsync,
MarkAllReadAsync) moves to a fragment file co-located with its feature folder.
The root partial in Infrastructure/ retains the constructor and implements
clause. Partial fragments deliberately keep the owning class's namespace
(Infrastructure) rather than matching folder; their *folder* identifies
the slice.
