+++ modules/Notifications/src/SimpleModule.Notifications/Features/Notifications/GetById/NotificationService.GetById.cs
@@ -1,0 +1,13 @@
+using Microsoft.EntityFrameworkCore;
+using SimpleModule.Notifications.Contracts;
+using SimpleModule.Users.Contracts;
+
+namespace SimpleModule.Notifications.Infrastructure;
+
+public sealed partial class NotificationService
+{
+    public async Task<Notification?> GetByIdAsync(NotificationId id, UserId userId) =>
+        await db
+            .Notifications.AsNoTracking()
+            .FirstOrDefaultAsync(n => n.Id == id && n.UserId == userId);
+}
