+++ modules/Notifications/src/SimpleModule.Notifications/Features/Notifications/MarkAllRead/NotificationService.MarkAllRead.cs
@@ -1,0 +1,15 @@
+using Microsoft.EntityFrameworkCore;
+using SimpleModule.Users.Contracts;
+
+namespace SimpleModule.Notifications.Infrastructure;
+
+public sealed partial class NotificationService
+{
+    public async Task<int> MarkAllReadAsync(UserId userId)
+    {
+        var now = DateTimeOffset.UtcNow;
+        return await db
+            .Notifications.Where(n => n.UserId == userId && n.ReadAt == null)
+            .ExecuteUpdateAsync(s => s.SetProperty(n => n.ReadAt, now));
+    }
+}
