+++ modules/Notifications/src/SimpleModule.Notifications/Features/Notifications/UnreadCount/NotificationService.UnreadCount.cs
@@ -1,0 +1,13 @@
+using Microsoft.EntityFrameworkCore;
+using SimpleModule.Users.Contracts;
+
+namespace SimpleModule.Notifications.Infrastructure;
+
+public sealed partial class NotificationService
+{
+    public Task<int> GetUnreadCountAsync(
+        UserId userId,
+        CancellationToken cancellationToken = default
+    ) =>
+        db.Notifications.CountAsync(n => n.UserId == userId && n.ReadAt == null, cancellationToken);
+}
